Adds loading component and restructures character.js into containers folder:
