+++ src/components/Loading/Loading.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Loading = () => {
+  return (
+    <div>
+      Server call is in progress ....
+    </div>
+  )
+}
+
+export default Loading
